tests/wpt/tests/html/webappapis/dynamic-markup-insertion/document-write/iframe_004.html

<!doctype html>
<title>document.write script into iframe write back into parent</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<iframe id="test"></iframe>
<script>
var t = async_test();
var iframe = document.getElementById("test");
var order = [];
t.step(function() {
  order.push(1);
  var s = "<script>parent.order.push(2); parent.document.write('<script>order.push(3);</script'+'>'); parent.order.push(4)</script" + ">";
  for (var i=0; i<s.length; i++) {
    iframe.contentDocument.write(s[i]);
  }
  iframe.contentDocument.close();
  order.push(5);
  assert_array_equals(order, [1,2,3,4,5])
}
);
t.done();
</script>
<div id="log"></div>
